separate SectionHeader class into its own file
